requirements.txt
dj-database-url==0.4.1
Django==1.9.7
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==2.0.6
appdirs==1.4.2
djangorestframework==3.5.4
packaging==16.8
pyparsing==2.1.10
six==1.10.0
pandas
bokeh